Recipe.Kitchen

Documentation

data Station

Representation of a station in a cooking environment, for example an oven.

Constructors

Station

Fields

stName :: StName

Name of the station, should be unique.

stConstrF :: ConstraintF

Constraint function (Recipe -> Maybe Process), given a recipe, if this station can perform the root action of the recipe then return Just a list of processes necessary to perform that action else return Nothing.

stObs :: [IO Obs]

Local observables for example temperature of an oven.

type StName = String

data Obs

An observable value.

Constructors

ObsTemp Int
Observable temperature.

ObsTime Time
Observable time.

ObsOpt String Bool
Observable option, labelled to correspond with a CondOpt in a recipe. True would mean to include that optional part of the recipe.

evalCond :: Condition -> [Obs] -> Bool

Given a list of observables, returns whether the given condition is true.

type ConstraintF = Recipe -> Maybe [Process]

popCond :: Recipe -> Recipe

Removes the Condition from the root action of a recipe. Does nothing if there is no condition.

addEvalCond :: Condition -> [Process] -> [Process]

Adds the EvalCond process after the Input process. If Input does not exist then EvalCond is added to the start.

popT :: Recipe -> Recipe

Removes the Transaction wrapper from the root action of the given recipe. Does nothing if there is no transaction wrapper.

data Process

A process that a station goes through, intended to be interfaced with device specific instructions.

Constructors

Input
Receive input.

Output
Output contents.

Preheat Int
Preheat to the given temperature.

DoNothing
Do nothing.

PCombine String
Combine contents with the given method.

EvalCond Condition
Evaluate the given condition.

MeasureOut Measurement
Measure our the given measurement of contents.

data Env

Representation of a cooking environment for example a kitchen.

Constructors

Env

Fields

eStations :: [Station]

List of stations present in the environment.

eObs :: [IO Obs]

Global observables for example time.