Recipe.Printer

Documentation

extractLabel :: [Tree (Label, Action)] -> [String]

Given a list of trees, extract either their labels or, if a tree's root node is an ingredient, extract the ingredient name.

extractLabel' :: Tree (Label, Action) -> String

Helper function for extractLabel. If a node contains a label and ingredient, return the name of the ingredient, else return the label as a String.

type Step = (Label, String)

Step is a pair of a label (step number) and the instructions for that step.

steps :: Recipe -> Tree Step

Translates a recipe into a tree of steps.

ppSteps :: Recipe -> IO ()

Prints one step per line in the following style. 1) ... 2) ...

ppTree :: Show a => Tree a -> IO ()

Prints the recipe as an ASCII tree.

ppIngredients :: Recipe -> IO ()

Prints one ingredient name per line.

ppIngredientsQ :: Recipe -> IO ()

Prints one ingredient name per line along with it's quantity.

printSchedule :: Schedule -> Map Label String -> Map Label Recipe -> IO ()

Print the given schedule using the given maps. The first is a map of step numbers to their instructions, the second is a map of labels to their recipes from the labelled tree.

printStack :: Stack -> Map Label String -> Map Label Recipe -> StateT Time IO ()

Print the given stack of actions. The first is a map of step numbers to their instructions, the second is a map of labels to their recipes from the labelled tree.

scheduleAndPrint :: Recipe -> Env -> IO ()

Schedule the given recipe in the given environemnt and print the resulting schedule.