Recipe.Properties

Documentation

type PropertyList v = [(String, v)]

List of String keys and properties of a given type.

lookupProp :: Monoid v => PropertyList v -> String -> v

Lookup the property of a given key in the PropertyList. Returns mempty if not found.

evalProps :: Monoid v => PropertyList v -> Recipe -> v

Cumulatively, evaluates the properties of every ingredient in the given recipe.

evalPropsQ :: (Integral v, Monoid v) => PropertyList (v, Measurement) -> [(String, Measurement)] -> v

The same as evalProps but considers the quantity of an ingredient used and calculates the ratio accordingly.

applyQuant :: Integral v => Measurement -> (v, Measurement) -> v

Apply the given measurement to the given quantified property. Returning the value of the property applicable to the given measurement. Informally: applyQuant 100g (£1,50g) means "Every 50g costs £1, I have 100g therefore that costs £2".

ratio :: Measurement -> Measurement -> Float

The ratio of two measurements.

ratio' :: Int -> Int -> Float

Helper function to avoid "divide by zero" in ratio.

newtype Price

Price is a wrapper around an Int representing pence.

Constructors

Price

Fields

pence :: Int

type PriceList = PropertyList (Price, Measurement)

PropertyList mapping an ingredient (String) to a pair of price per measurement.

price :: PriceList -> Recipe -> Price

Given a price list, calculate the price of a recipe.

type Cuisine a = PropertyList a -> Recipe

data Seasoning

Representation of the Salt, Sweet, Acid seasoning triangle.

Constructors

Salt

Sweet

Acid

type Profile = Map Seasoning Int

A profile is a mapping of a seasoning to the number of times it occurs.

profile :: [Seasoning] -> Profile

Given a list of seasonings, produce a profile of their frequencies.

balance :: Recipe -> PropertyList Seasoning -> [(Seasoning, Int)]

Given a recipe and a property list of seasonings, find the seasoning that each ingredient offers. Then profile said seasonings and return a list showing how much each seasoning needs to be increased by to balance the recipe. For example if a recipe had a profile of [(Salt, 1), (Sweet, 0), (Acid, 2)] then balance would return [(Salt, 1), (Sweet, 2), (Acid, 0)].