Recipe.QS

Description

Contains QuickCheck properties for testing and code necessary to run QuickSpec on recipes.

Documentation

genIng :: Gen Recipe

Generate an ingredient. Name is random between "1" and "100".

genUnRec :: Gen Recipe -> Gen Recipe

Generate a unary recipe.

genBinRec :: Gen Recipe -> Gen Recipe -> Gen Recipe

Generate a binary recipe.

genTemp :: Gen Int

Generate a temperature between 100 and 240.

genMethod :: Gen String

Generate a combine method: "mix", "spread", "wrap".

singleCond :: Gen Condition

Generate a single condition (not AND or OR).

mkUsefulEnv :: Env

Make an environment containing 3 useful stations and 1 useless station.

mkUselessEnv :: Env

Make an environement containing 2 useless stations.

timeZero :: IO Obs

ObsTime 0

mkStName :: Int -> StName

mkStName 1 = Station_1

mkStationUseful :: StName -> Station

Make a station with the given name that can do everything.

mkStationUseless :: StName -> Station

Make a station with the given name that can't do anything.

qsRecipe :: IO ()

mkProp :: Condition -> Condition -> [Obs] -> Bool

Takes two conditions and compares the result of running evalCond on each of them with the given observables.

prop_and_comm :: Condition -> Condition -> [Obs] -> Bool

prop_and_id :: Condition -> [Obs] -> Bool

prop_or_comm :: Condition -> Condition -> [Obs] -> Bool

prop_or_id :: Condition -> [Obs] -> Bool

prop_and_assoc :: Condition -> Condition -> Condition -> [Obs] -> Bool

prop_and_or_id :: Condition -> Condition -> [Obs] -> Bool

prop_or_and_id :: Condition -> Condition -> [Obs] -> Bool

prop_or_assoc :: Condition -> Condition -> Condition -> [Obs] -> Bool

prop_quickspec_law_11 :: Condition -> Condition -> Condition -> [Obs] -> Bool

prop_and_distr_and :: Condition -> Condition -> Condition -> [Obs] -> Bool

prop_and_distr_or :: Condition -> Condition -> Condition -> [Obs] -> Bool

prop_or_distr_or :: Condition -> Condition -> Condition -> [Obs] -> Bool

prop_or_distr_and :: Condition -> Condition -> Condition -> [Obs] -> Bool

prop_min_hours :: Time -> Bool

prop_min_min :: Time -> Bool

prop_combine_comm :: String -> Recipe -> Recipe -> Bool

prop_labelR_root :: Recipe -> Bool

prop_labelA_unlabel :: Tree Action -> Bool

prop_fold_time :: Tree Action -> Bool

prop_fold_ing :: Tree Action -> Bool

prop_eval_cond_false_empty :: Condition -> Bool

prop_eval_cond_false :: Condition -> [Obs] -> Bool

failObs :: Condition -> [Obs] -> [Obs]

Generate observables for which a condition will fail.

prop_eval_cond_true :: Condition -> [Obs] -> Bool

condToObs :: Condition -> [Obs]

Generate observables for a condition to evalCond to True.

prop_scheduling_succeed :: Recipe -> Bool

prop_scheduling_fail :: Recipe -> Property

getAllLabels :: Schedule Label -> [Label]

runTests :: IO Bool