Recipe.Recipe

Documentation

type Recipe = Tree Action

Recipe type is simply a Data.Tree of Actions.

data Action

Actions represent something that is done during the recipe and are taken to be "applied" on their child nodes in the recipe thus providing sequencing. Not used directly, used via the helper functions.

Constructors

GetIngredient String

Heat

HeatAt Int

Wait

Combine String

Conditional Action Condition

Transaction Action

Measure Measurement

newtype Time

Time is a wrapper around an Int representing seconds.

Constructors

Time Int

data Condition

Conditions represent an event which you perform a recipe "until".

Constructors

CondTime Time
Until the given time has elapsed.

CondTemp Int
Until the temperature has been reached.

CondOpt String
Optional step, labelled for identification. See observables (Obs).

Condition `AND` Condition
Logical "and" of two conditions.

Condition `OR` Condition
Logical "or" of two conditions.

data Measurement

Represents a measurement of something.

Constructors

Count Int
Number of something e.g. 1 apple.

Grams Int
Number of grams.

Milliletres Int
Number of milliletres.

getMeasure :: Measurement -> Int

Returns the magnitude of a Measurement.

ingredient :: String -> Recipe

Get an ingredient with the given name.

heat :: Recipe -> Recipe

Heat the given recipe, used when the temperature it is heated at doesn't matter for example heating water in a kettle.

heatAt :: Int -> Recipe -> Recipe

Heat the given recipe at the given temperature. (HeatAt).

wait :: Recipe -> Recipe

Wait after performing the given recipe. Without being wrapped with CondTime is an infinitely small wait.

combine :: String -> Recipe -> Recipe -> Recipe

Combine the two recipes using the method given as a String e.g. "mix".

addCondition :: Condition -> Recipe -> Recipe

Add the given condition to the root action of the given recipe. If the action is already wrapped with a condtion, (AND) is applied to the two conditions.

transaction :: Recipe -> Recipe

Wraps the root node of a recipe in a Transaction.

measure :: Measurement -> Recipe -> Recipe

Adds a Measure action with the given Measurement above the given recipe.

optional :: String -> Recipe -> Recipe

Applies addCondition CondOpt with the given label.

toTemp :: Int -> Recipe -> Recipe

Applies addCondition CondTemp with the given temperature.

forTime :: Time -> Recipe -> Recipe

Applies addCondition CondTime with the given time.

(.&&) :: Condition -> Condition -> Condition

Creates an AND condition.

(.||) :: Condition -> Condition -> Condition

Creates an OR condition.

hours :: Time -> Time

Creates a time with the given number of hours: hours 1 = Time 3600.

minutes :: Time -> Time

Creates a time with the given number of minutes: minutes 1 = Time 60.

foldCond :: (Ord a, Monoid a) => (Condition -> a) -> Condition -> a

Folds a function over a condition. AND will mappend the values from folding over the two conditions whereas OR takes the max.

foldRecipe :: Monoid a => (Action -> a) -> Recipe -> a

Fold a function over a recipe to obtain a value.

ingredients :: Recipe -> [String]

The ingredients of a recipe.

ingredientsQ :: Recipe -> [(String, Measurement)]

List of ingredients paired with their measurements. If no measurement is present for an ingredient, it is paired with "Count 0".

type Label = Int

labelRecipe :: Recipe -> Tree Label

Replaces the action in each node with a unique label.

labelRecipeR :: Recipe -> Tree (Label, Recipe)

Labels each node of the tree while keeping the entire recipe from that node downwards in a tuple.

labelRecipeA :: Recipe -> Tree (Label, Action)

Same as labelRecipeR but only keeps the action from that node.

tempToTime :: Int -> Time

Time to reach a certain temperature, for use with CondTemp.

preheatTime :: Int -> Time

Time taken to preheat to a given temperature, for use with heatAt.

time :: Recipe -> Time

Estimate of the time taken to execute a recipe using timeAction.

timeAction :: Action -> Time

Estimate of the time taken to perform a certain action.

topologicals :: Recipe -> [[Action]]

Returns a list of all the possible topological sorts of a recipe.

isLeaf :: Recipe -> Bool

Returns True if the given node is a leaf i.e. has no child nodes.

leaves :: Recipe -> [Recipe]

List of all leaves in the recipe.

removeFrom :: Recipe -> Recipe -> Recipe

Removes all occurences of a sub recipe from the given recipe. Removing a recipe from itself does nothing.

deleteAll :: Recipe -> [Recipe] -> [Recipe]

Removes all occurences of a given recipe in a list of recipes.