Recipe.Scheduler

Description

The scheduling function implemented here is mainly for demonstration purposes. It is rather inefficient and runs for a very long time thus it is not suitable for scheduling recipes with more than a small number of Actions.

Leaf Selection Heuristics

Shortest task.

Longest branch.

Stack Selection Heuristics

Least in-demand first. Calculates an expected time that will be added to the stack by all the unscheduled tasks.

Least idle time required i.e. the stack which has a height closest to the minimum start time of the action being scheduled.

Smallest stack.

Types

data Task a

Task performed by a Station.

Constructors

Active a
Active performing the action corresponding to the label.

Idle Time
Idle for a time.

type Stack a = [Task a]

type Schedule a = Map StName (Stack a)

Schedule is a map of station names to their Stack of Tasks

Label Helper Functions

mkLabelMap :: Tree (Label, Recipe) -> Map Label Recipe

Create a map of labels to their recipe in the given tree.

lookupR :: Label -> Map Label Recipe -> Recipe

Given a label, lookup the corresponding recipe in the map.

childLabels :: Label -> Tree Label -> [Label]

Get the child labels of a given label in a given tree.

lookupValSts :: Env -> Label -> Map Label Recipe -> [StName]

Lookup the stations capable of handling the action at the root node of the recipe corresponding the to given label.

validStations :: Env -> Recipe -> [StName]

Lookup the stations capable of handling the action at the root node of the given recipe.

Scheduling Functions

duration :: Label -> Map Label Recipe -> Time

Duration that the root action of the recipe corresponding to the given label will take. Uses timeAction.

scheduleRecipe :: Recipe -> Env -> Schedule Label

Schedules a recipe in the given environment. Will error if no schedule is available e.g. if there is no Station capable of handling a certain Action. See documentation at the top for heuristics used.

initSchedule :: Env -> Schedule Label

Creates a schedule where each Station in the given environment has an empty Stack.

schLength :: Schedule Label -> Recipe -> Time

Time a schedule takes, must be given the recipe that the schedule is for.