Recipe.Simulator

Description

Allows the simulation of a recipe in the given environment, printing progress to sdout.

Documentation

simulate :: Recipe -> Env -> Float -> Bool -> IO ()

Schedule the given recipe in the given environment then simulate it with the given speed multiplier. Also takes a Bool to determine whether to clear stdout after each round of tasks (True means clear).